requirements/dev.txt
# requirements/dev.txt
# run pip install -r requirements/dev.txt to load development requirments specific to development virtual envs
-r base.txt
django-debug-toolbar==1.0.1